illustration page fixes
