--- src/layouts/IllustrationLayout.tsx
@@ -1,11 +1,0 @@
-import styles from '@/styles/ProjectLayout.module.css';
-
-const IllustrationLayout = (
-    {children}: {children: React.ReactNode}
-    ) => {
-    return <div className={styles.projectGallery}>
-        {children}  
-    </div>
-};
-
-export default IllustrationLayout;
