+++ src/layouts/ProjectLayout.tsx
@@ -1,11 +1,32 @@
 import styles from '@/styles/ProjectLayout.module.css';
+import phfont from '@/utils/phfont';
+
+interface projectInfoInterface {
+    title: string,
+    id: string
+}
 
 const ProjectLayout = (
-    {children}: {children: React.ReactNode}
+    {
+        children, pageInfo
+    }: {children: React.ReactNode, pageInfo?: projectInfoInterface}
     ) => {
-    return <div className={styles.projectGallery}>
-        {children}  
-    </div>
+
+        const projectTitleDisplay = () => {
+            return pageInfo ? <div className={styles.projectSubtitle}>
+            <h2 id={pageInfo.id} className={`${styles.projectSubtitleTag} ${phfont.className}`}>{pageInfo.title}</h2>
+            </div>
+            : null
+        }
+
+    return (
+        <div>
+            {projectTitleDisplay()}
+            <div className={styles.projectGallery}>
+                {children}  
+            </div>
+        </div>
+    )
 };
 
 export default ProjectLayout;
