+++ src/utils/phfont.tsx
@@ -1,0 +1,8 @@
+import { Patrick_Hand } from "next/font/google";
+
+const phfont = Patrick_Hand({
+    weight: ["400"],
+    subsets: ['latin'],
+  })
+
+  export default phfont;
